Split sharding.py into separate files
